requirements.txt
aiohttp
python-binance
pandas
numpy
ta
gspread
oauth2client
python-dotenv
flask
